requirements.txt
<<<<<<< HEAD
flatten_dict==0.4.2
lxml==6.0.2
Flask==3.1.2 
Flask-SocketIO==5.5.1
=======
flatten_dict==0.4.2
lief==0.17.1
lxml==6.0.2
pikepdf==10.0.2
pillow==12.0.0
pyelftools==0.32
>>>>>>> 9bd8813337e6acd175f40eb3a3ae26539b1cf83c
